refactor: separate targets with .native ext
