--- packages/fiber/native/package.json
@@ -1,5 +1,0 @@
-{
-  "main": "dist/react-three-fiber-native.cjs.js",
-  "module": "dist/react-three-fiber-native.esm.js",
-  "types": "dist/react-three-fiber-native.cjs.d.ts"
-}
