--- packages/fiber/src/native.tsx
@@ -1,17 +1,0 @@
-export * from './three-types'
-import * as ReactThreeFiber from './three-types'
-export { ReactThreeFiber }
-export type {
-  Intersection,
-  Subscription,
-  Dpr,
-  Size,
-  Viewport,
-  Camera,
-  RenderCallback,
-  Performance,
-  RootState,
-} from './core/store'
-export type { ThreeEvent, Events, EventManager } from './core/events'
-export * from './native/Canvas'
-export * from './native/index'
